Add protectors to wrap fscrypt keys
Currently the raw fscrypt key (aka policy key) is wrapped with a key
generated from the user password.

With this change, policy keys are wrapped with an intermediate key
called a protector, and it's the protector that needs the user
password.

This allows having a protector for multiple policies (one password can
decrypt N directories) and multiple protectors for the same policy (N
valid passwords for the same directory).

This approach is taken from https://github.com/google/fscrypt