Check if a protector exists before asking for its password
In commands like unlock or change-password the user can indicate the
ID of the specific protector to use, but until now the code assumed
that the protector existed and asked for its password.

With this change those commands fail immediately if the protector does
not exist or is not used in that directory.