Sort result of get_protectors_for_policy()
This ensures that protectors are always returned in the same order. We
also give priority to TPM2 protectors over password protectors since
TPM PINs are likely to be shorter and therefore expected to be the
preferred authentication method.

If this is not enough we can add later a way to set the priority of
the different protectors.