Use the correct credential name when creating or updating a protector
The code uses 'password' all over the place for the secret that the
user enters to unlock a protector, but for user-visible strings let's
use a credential name specific to the protector, like "password" or
"TPM2 PIN".

We were already using Protector::get_prompt() for existing secrets,
this patch changes the prompt for new secrets.