PAM: unwrap the protector key only once
During a normal PAM session we need to unwrap the protector key twice:
once in authenticate() (to authenticate the user) and then again in
open_session() (to actually unlock the home directory).

This is generally not a problem apart from the added overhead of
unwrapping the key twice (which can be costly depending on the KDF and
other factors), but in the case of FIDO2 tokens this requires us to
confirm presence (e.g physically touch the token) twice in a row,
which is unexpected and annoying.

Instead, we can keep the protector key that we got during
authentication and use it directly during open_session() without
having to unwrap it again.