Add Config::check()
This method checks for errors in the config file. All other methods
that get configuration values will now panic if there is an error.

With this, users are expected to run dirlock::init() early on and bail
if there is a problem. After that, all other uses of the configuration
are guaranteed to succeed.