fscrypt: don't return an error if a filesystem does not support encryption
The FS_IOC_GET_ENCRYPTION_POLICY_EX ioctl returns ENOTTY if encryption
is not supported by that filesystem (vfat) and EOPNOTSUPP if it's
supported but not enabled in the superblock (ext4).

In both cases this simply means "this directory is not encrypted" and
is a common, expected situation.

However, it is also possible to get those return values if the
directory *is* encrypted but the kernel is too old or does not have
encryption enabled. So in order to tell what's really going on we use
statx(3) and check the STATX_ATTR_ENCRYPTED flag.