Use the TPM PIN to encrypt the protector key
When we create a Tpm2Protector we generate a random protector key
and seal it with the TPM, using a value derived from the PIN for
authorization.

This patch adds an additional layer of protection by encrypting the
protector key with an AES key, also derived from the PIN but different
from the TPM authorization value.