dirlockd: Add a ConvertDir method to encrypt a directory with data
This method allows converting an existing dir into an encrypted one in
place. This is an asynchronous method, and comes with a few related
methods and signals to cancel the background job and query its status.

This should probably be merged with EncryptDir, but let's keep it
separate for now.